packages/simcore_ai/pyproject.toml
[project]
name = "simcore-ai"
version = "0.2.0"
description = "Provider-agnostic AI engine for SimWorks (pure Python)."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
  { name = "Jackfruit Co SimWorks" },
  { name = "Tyler J", email = "tyler@jackfruit.me" }
]
classifiers = [
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
]

dependencies = [
  "pydantic>=2.8",
  "typing-extensions>=4.12",
  "opentelemetry-api>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6", "mypy>=1.11"]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simcore_ai"]

[tool.hatch.build.targets.sdist]
include = [
  "src/simcore_ai/**",
  "pyproject.toml",
  "README.md",
  "LICENSE*",
]